web funcionando v2

+++ src/main/java/cburgosdev/java/Constants/StoreConstants.java
@@ -1,5 +1,5 @@
 package cburgosdev.java.Constants;
 
 public class StoreConstants {
-    public static int Ripley = 1;
+    public static Long Ripley = 1L;
 }
