+++ src/main/java/cburgosdev/java/Models/ProductDetail.java
@@ -9,26 +9,26 @@ import java.util.Date;
 public class ProductDetail {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    private int productId;
+    private Long id;
+    private Long productId;
     private Double price;
     private Double priceWithCard;
     private Double minPrice;
     private Date date;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
-    public void setProductId(int productId) {
+    public void setProductId(Long productId) {
         this.productId = productId;
     }
