+++ src/main/java/cburgosdev/java/Services/IProductService.java
@@ -5,5 +5,5 @@ import cburgosdev.java.DTOs.ProductDTO;
 import java.util.HashMap;
 
 public interface IProductService {
-    HashMap<String, Integer> saveProductAndDetail(ProductDTO productBean, int storeId, int categoryId);
+    HashMap<String, Integer> saveProductAndDetail(ProductDTO productBean, Long storeId, Long categoryId);
 }
